+++ packages/blade/.storybook/react-native/storybook.requires.js
@@ -28,6 +28,7 @@ const getStories = () => {
     require('../../src/components/Button/Button/Button.stories.tsx'),
     require('../../src/components/Checkbox/Checkbox.stories.tsx'),
     require('../../src/components/Checkbox/CheckboxGroup.stories.tsx'),
+    require('../../src/components/Counter/Counter.stories.tsx'),
     require('../../src/components/Icons/Icons.stories.tsx'),
     require('../../src/components/Indicator/Indicator.stories.tsx'),
     require('../../src/components/Input/BaseInput/BaseInput.stories.tsx'),