+++ packages/blade/src/components/Input/PasswordInput/PasswordInput.stories.tsx
@@ -2,11 +2,12 @@ import type { ComponentStory, Meta } from '@storybook/react';
 import { Title, Subtitle, Primary, ArgsTable, Stories, PRIMARY_STORY } from '@storybook/addon-docs';
 import type { ReactElement } from 'react';
 import { useState } from 'react';
-import { Highlight, Link } from '@storybook/design-system';
 
 import type { PasswordInputProps } from './PasswordInput';
 import { PasswordInput } from './PasswordInput';
 import useMakeFigmaURL from '~src/_helpers/storybook/useMakeFigmaURL';
+import Sandbox from '~src/_helpers/storybook/Sandbox';
+import FigmaEmbed from '~src/_helpers/storybook/FigmaEmbed';
 
 const Page = (): ReactElement => {
   const figmaURL = useMakeFigmaURL([
@@ -35,10 +36,23 @@ const Page = (): ReactElement => {
         using an optional reveal button.
       </Subtitle>
       <Title>Usage</Title>
-      <Link withArrow={true} href={figmaURL} target="_blank" rel="noreferrer noopener">
-        View in Figma
-      </Link>
-      <Highlight language="tsx">{`import { PasswordInput } from '@razorpay/blade/components' \nimport type { PasswordInputProps } from '@razorpay/blade/components'`}</Highlight>
+      <FigmaEmbed title="Password Figma Designs" src={figmaURL} />
+      <Sandbox showConsole>
+        {`
+          import { PasswordInput } from '@razorpay/blade/components';
+
+          function App(): JSX.Element {
+            return (
+              <PasswordInput 
+                label="Enter Password" 
+                onChange={(e) => console.log(e)} 
+              />
+            )
+          }
+
+          export default App;
+        `}
+      </Sandbox>
       <Title>Example</Title>
       <Subtitle>You can change the properties using the controls in the table below.</Subtitle>
       <Primary />