+++ packages/blade/src/components/Spinner/Spinner/Spinner.stories.tsx
@@ -1,12 +1,13 @@
 import type { ComponentStory, Meta } from '@storybook/react';
 import { Title, Subtitle, Primary, ArgsTable, Stories, PRIMARY_STORY } from '@storybook/addon-docs';
-import { Highlight, Link } from '@storybook/design-system';
 import type { ReactElement } from 'react';
 import type { SpinnerProps } from './Spinner';
 import { Spinner as SpinnerComponent } from './Spinner';
 import useMakeFigmaURL from '~src/_helpers/storybook/useMakeFigmaURL';
 import Box from '~components/Box';
 import { Text } from '~components/Typography';
+import Sandbox from '~src/_helpers/storybook/Sandbox';
+import FigmaEmbed from '~src/_helpers/storybook/FigmaEmbed';
 
 const Page = (): ReactElement => {
   const figmaURL = useMakeFigmaURL([
@@ -32,13 +33,32 @@ const Page = (): ReactElement => {
       <Subtitle>
         A spinner is an element with a looping animation that indicates loading is in process.
       </Subtitle>
-      <Link withArrow={true} href={figmaURL} target="_blank" rel="noreferrer noopener">
-        View in Figma
-      </Link>
+      <FigmaEmbed src={figmaURL} title="Spinner Figma Designs" />
       <br />
       <br />
       <Title>Usage</Title>
-      <Highlight language="tsx">{`import { Spinner } from '@razorpay/blade/components'; \nimport type { SpinnerProps } from '@razorpay/blade/components';`}</Highlight>
+      <Sandbox>
+        {`
+          import { useEffect, useState } from 'react';
+          import { Spinner, Text } from '@razorpay/blade/components';
+
+          function App(): JSX.Element {
+            const [isLoading, setIsLoading] = useState(true);
+
+            useEffect(() => {
+              setTimeout(() => {
+                setIsLoading(false)
+              }, 5000)
+            }, [])
+
+            return (
+              isLoading ? <Spinner /> : <Text>Tadaa 🥳 Reload sandbox to see spinner again</Text>
+            )
+          }
+
+          export default App;
+        `}
+      </Sandbox>
       <Title>Example</Title>
       <Subtitle>
         This is the default spinner. You can change the properties of this spinner using the